+++ ios/Exponent/Kernel/AppLoader/AppFetcher/EXAppFetcher+Private.h
@@ -2,7 +2,8 @@
 
 #import <Foundation/Foundation.h>
 #import "EXAppFetcher.h"
-#import <EXManifests/EXManifestsManifest.h>
+
+@class EXManifestsManifest;
 
 NS_ASSUME_NONNULL_BEGIN