+++ ios/Exponent/Kernel/Views/Loading/EXManagedAppSplashScreenConfigurationBuilder.h
@@ -1,6 +1,7 @@
 #import <Foundation/Foundation.h>
 #import "EXManagedAppSplashScreenConfiguration.h"
-#import <EXManifests/EXManifestsManifest.h>
+
+@class EXManifestsManifest;
 
 NS_ASSUME_NONNULL_BEGIN