+++ ios/Exponent/Versioned/Core/EXVersionManager.h
@@ -2,7 +2,8 @@
 
 #import <Foundation/Foundation.h>
 #import <React/RCTLog.h>
-#import <EXManifests/EXManifestsManifest.h>
+
+@class EXManifestsManifest;
 
 @interface EXVersionManager : NSObject